feat: drop legacy titolo column from encik schema
The titolo column was a backward-compat shim from autish-legacy. All display and search code now uses terminologio dict as source of truth.

Changes:
- Removed titolo from _CREATE_ENCIK DDL (TEXT NOT NULL DEFAULT '')
- Added migration: backfill terminologio from titolo for old entries, then ALTER TABLE DROP COLUMN titolo (graceful fallback for SQLite < 3.35)
- row_to_dict now synthesizes entry['titolo'] from terminologio for display backward compat (not from DB column)
- service.create() no longer populates titolo column
- migrate_from_autish.py backfills terminologio inline, drops titolo from INSERT
- _time_entry.py: removed 'titolo' key from data dicts

Also split monolith files >500 lines:
- service.py (592->243): extracted SearchMixin into _search_service.py
- display.py (637->24): extracted _display_html, _display_entry, _display_panel, _display_graph modules

Closes #28
